+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # vdlx-datagrid changelog
 
+## v4.6.2
+
+- Export data as CSV: New vdlx-datagrid attributes `show-export` and `export-filename` that display a datagrid header 
+  button to allow the user to download the current table contents to a CSV file.
+- Save any pending cell edits when the view is unloaded. 
+
 ## v4.6.1
 
 - Added debug bundle of the vdlx-datagrid library. This can be used in place of the standard library bundle to facilitate